A ton of new stuff was added!
Added a "goal" for the player to reach.

Added a customized ground for the levels that can be drawn in the editor. And attached script will then create an edge collider along the drawn ground to make it walkable! Additionally, it can wiggle, bend, be redrawn during gameplay, and is easily extensible!

Added a ton of great-looking post-processing effects and feedbacks to make the game feel more alive.